feat(sources): add /sources page for non-chat source management
Closes #673. Information sources (RSS, GitHub releases/issues, arXiv)
could only be registered via the chat `manageSource` tool. Backend
API was complete; only a non-chat UI entry point was missing.

Extract the sources-manager UI from manageSource/View.vue into a
reusable SourcesManager component that owns its own state via
apiGet. Both contexts now mount it:

- Plugin View (chat tool result) passes the tool-result data as an
  initial seed via the `initialData` prop.
- /sources page (new) mounts it without a seed; SourcesManager
  fetches from /api/sources on mount.

A watch on `initialData` handles the plugin case where the user
switches between manageSource tool results without remounting.

The old 718-line View.vue is now a 12-line wrapper. All the form
validation, preset installation, rebuild / brief-loading logic
lives in the shared component.

Nav entry added to PluginLauncher between Wiki and Skills (data-
plugin side of the separator, fitting alongside todos/scheduler/
wiki). i18n keys for all 8 locales. No ⌘ shortcut — existing
⌘3-⌘8 are already densely packed and the tooltip drift would be
painful to expand.
