+++ src/components/SourcesView.vue
@@ -1,0 +1,7 @@
+<template>
+  <SourcesManager />
+</template>
+
+<script setup lang="ts">
+import SourcesManager from "./SourcesManager.vue";
+</script>
