+++ src/lang/de.ts
@@ -155,6 +155,7 @@ const deMessages = {
     todos: { label: "To-dos", title: "To-dos öffnen (⌘4)" },
     scheduler: { label: "Zeitplan", title: "Zeitplan öffnen (⌘5)" },
     wiki: { label: "Wiki", title: "Wiki öffnen (⌘6)" },
+    sources: { label: "Quellen", title: "Informationsquellen öffnen" },
     skills: { label: "Skills", title: "Skills öffnen (⌘7)" },
     roles: { label: "Rollen", title: "Rollen öffnen (⌘8)" },
     files: { label: "Dateien", title: "Workspace-Dateien öffnen (⌘3)" },
