+++ src/lang/en.ts
@@ -176,6 +176,7 @@ const enMessages = {
     todos: { label: "Todos", title: "Open todos (⌘4)" },
     scheduler: { label: "Schedule", title: "Open schedule (⌘5)" },
     wiki: { label: "Wiki", title: "Open wiki (⌘6)" },
+    sources: { label: "Sources", title: "Open information sources" },
     skills: { label: "Skills", title: "Open skills (⌘7)" },
     roles: { label: "Roles", title: "Open roles (⌘8)" },
     files: { label: "Files", title: "Open workspace files (⌘3)" },
